Improved article loading functionality
Rather than one call to read an entire db - two separate calls are now
made. One indicating the category and which index to load from. To the
mindset behind this is that it will only load 20 articles at a time -
and then this method can be re-called when additional articles are
needed.